feat: add policies - only log in user get their tasks

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Http\Resources\TaskResource;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -15,7 +16,9 @@ class TaskController extends Controller
      */
     public function index()
     {
-        return TaskResource::collection(auth()->user()->tasks()->get());
+        Gate::authorize('viewAny', Task::class);
+
+        return TaskResource::collection(Task::all());
     }
 
     /**
@@ -23,7 +26,12 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
-        $task = $request->user()->tasks()->create($request->validated());
+
+        if(request()->user()->cannot('create', Task::class)) {
+            abort(403, "This action is unauthorized");
+        }
+
+        $task = Task::create($request->validated());
 
         return TaskResource::make($task);
     }
@@ -33,6 +41,7 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        Gate::authorize('view', $task);
         return TaskResource::make($task);
     }
 
@@ -41,6 +50,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if($request->user()->cannot('update', $task )) {
+            abort(403, "This action is unauthorized");
+        }
+
         $task->update($request->validated());
 
         return TaskResource::make($task);
@@ -51,6 +64,10 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        if(request()->user()->cannot('delete', $task )) {
+            abort(403, "This action is unauthorized");
+        }
+
         $task->delete();
 
         return response()->noContent();
